Requesting blocks

+++ wire/message.go
@@ -31,11 +31,11 @@ func MarshalMessage(msg *Message) []byte {
 	var buf bytes.Buffer
 
 	if msg.KeepAlive {
-		binary.Write(&buf, binary.BigEndian, uint64(0))
+		binary.Write(&buf, binary.BigEndian, uint32(0))
 		return buf.Bytes()
 	}
 
-	binary.Write(&buf, binary.BigEndian, 1+len(msg.Payload))
+	binary.Write(&buf, binary.BigEndian, uint32(1+len(msg.Payload)))
 	buf.WriteByte(byte(msg.ID))
 	buf.Write(msg.Payload)
